Main nuovo

+++ FSTSP/src/fstsp/NodesToUpdate.java
@@ -2,13 +2,13 @@ package fstsp;
 
 public class NodesToUpdate {
 
-	public int maxSavings;
+	public double maxSavings;
 	public Node iStar;
 	public Node jStar;
 	public Node kStar;
 
 	public NodesToUpdate() {
-		maxSavings = 0;
+		maxSavings = 0.0;
 		iStar = jStar = kStar = null;
 	}
 }
